test(session): move delegation tool tests to part 04
The two delegation-tool tests covering superseded instance release and
byte-stable tool blocks were relocated out of part 01 into a new part 04
file, leaving part 01 with only the tests that remain there.

Auto-committed-on: macbook
